Fix my_sleep_msecs argument type
The function prototype of my_sleep_msecs (in my.h) accepts an unsigned int
while its definition (in my.c) accepts an unsigned long int.

The function is only called from tui_sys_sleep which accepts an unsigned
long int, so the intended type seems to be unsigned long int.

Thus, the prototype is changed to reflect the intended type.

+++ src/my.h
@@ -98,7 +98,7 @@ enum {	/* unbufferd i/o related constants & masks */
 #ifndef MY_C
 extern int my_cursor_onoff( int onoff );
 extern int my_getch( unsigned int *outKeyMask );
-extern int my_sleep_msecs( unsigned int msecs );
+extern int my_sleep_msecs( unsigned long int msecs );
 extern int my_cls( void );
 extern int my_console_width( void );
 extern int my_console_height( void );
